like totaly cast doubles to floats
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10496 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/FontManager.cxx
@@ -529,7 +529,7 @@ ImageFont*    FontManager::getClosestRealSize(int faceID, float desiredSize, flo
 
 void	    FontManager::getPulseColor(const GLfloat *color, GLfloat *pulseColor) const
 {
-  float pulseTime = TimeKeeper::getCurrent().getSeconds();
+  float pulseTime = (float)TimeKeeper::getCurrent().getSeconds();
 
   // depth is how dark it should get (1.0 is to black)
   float pulseDepth = BZDBCache::pulseDepth;
